refactor(web): merge biomarker overview hero with About + 2-col experiments
User feedback: top of biomarker page should be one section (title +
description + the 3 About columns), and the experiment list should
be 2 columns instead of 1.

Layout-client now hosts only breadcrumb + tabs + private trend slot.
The Overview tab owns its full hero: eyebrow + title + summary + 3
About columns on the left, private trend card on the right, all in
one grid. The Research tab gets its own inline title + 1-line subtitle
since the layout no longer renders the page title.

Experiment cards drop from a single column to a 2-column grid on lg+
breakpoints; mobile stays 1-column. The standalone BiomarkerHeader
and BiomarkerAbout components are deleted because their content is
now inline in BiomarkerOverview and BiomarkerResearch.

Verification: pnpm typecheck, pnpm lint clean. Vitest 21/21 across
the three biomarker test files. Visual readback at
/biomarkers/resting-heart-rate (Overview + Research).
